Added easier setup
The config file no longer has to be edited manually and is, instead,
created if its not existing. User will be prompted to input information
and set a working directory.

Working directory stores the compiled files and traces.

This change allows users to call the cli from anywhere if they correctly
set the --configfile option and compiles the source files as necessary.