Restructured for testing with Protostar
I want to use protostart for testing and this commit already includes a
simple test file test_validate.cairo. For now the output builtin in the
respective file has to be deleted before using protostar to test it.

Asked if there is a better way - if not I can create a small script to
remove it and readd it after testing.

cairo/protostar.toml might get removed

test from inside the cairo dir with `protostar test ./tests
--cairo-path=./src`

README updates will follow when this is usable